requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.2.3
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.3
frozenlist==1.3.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
openai==0.27.2
python-dotenv==1.0.0
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.6
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
yarl==1.8.2
